Split opening the browser, and exchanging code into two
This makes the flow easier to follow, and allows us to split each part into separate methods (next commit).

Also removing the `currentAuthorizationSession` variable. We only need it if we were using universal links, and calling `resumeExternalUserAgentFlowWithURL` when an app link is received. Since we are using `ASWebAuthenticationSession`, we don't need it.